// applications/logger_fe/src/basic_data/proto/com_msg.ts
// @generated by protobuf-ts 2.8.1
// @generated from protobuf file "com_msg.proto" (package "ld_data", syntax proto3)
// tslint:disable
/* eslint-disable */
// @ts-nocheck
//
// 定义消息传输中的 protobuf 格式.
//
//传输的消息可以为字符串也可以为二进制数据.
//如果是字符串的话, 数据放在 str_data 字段中. 一般为 json 格式.
//如果是二进制数据的话, 数据放在 bin_data 字段中. 一般为另外一个序列化的 protobuf 数据.
//
import type { BinaryWriteOptions } from '@protobuf-ts/runtime'
import type { IBinaryWriter } from '@protobuf-ts/runtime'
import { WireType } from '@protobuf-ts/runtime'
import type { BinaryReadOptions } from '@protobuf-ts/runtime'
import type { IBinaryReader } from '@protobuf-ts/runtime'
import { UnknownFieldHandler } from '@protobuf-ts/runtime'
import type { PartialMessage } from '@protobuf-ts/runtime'
import { reflectionMergePartial } from '@protobuf-ts/runtime'
import { MESSAGE_TYPE } from '@protobuf-ts/runtime'
import { MessageType } from '@protobuf-ts/runtime'
/**
 * @generated from protobuf message ld_data.ComMsg
 */
export interface ComMsg {
  /**
   * @generated from protobuf field: string msg = 1;
   */
  msg: string // 标识传输消息的名称.
  /**
   * @generated from protobuf field: string str_data = 2;
   */
  strData: string
  /**
   * @generated from protobuf field: bytes bin_data = 3;
   */
  binData: Uint8Array
}
// @generated message type with reflection information, may provide speed optimized methods
class ComMsg$Type extends MessageType<ComMsg> {
  constructor() {
    super('ld_data.ComMsg', [
      { no: 1, name: 'msg', kind: 'scalar', T: 9 /*ScalarType.STRING*/ },
      { no: 2, name: 'str_data', kind: 'scalar', T: 9 /*ScalarType.STRING*/ },
      { no: 3, name: 'bin_data', kind: 'scalar', T: 12 /*ScalarType.BYTES*/ },
    ])
  }
  create(value?: PartialMessage<ComMsg>): ComMsg {
    const message = { msg: '', strData: '', binData: new Uint8Array(0) }
    globalThis.Object.defineProperty(message, MESSAGE_TYPE, { enumerable: false, value: this })
    if (value !== undefined) reflectionMergePartial<ComMsg>(this, message, value)
    return message
  }
  internalBinaryRead(
    reader: IBinaryReader,
    length: number,
    options: BinaryReadOptions,
    target?: ComMsg
  ): ComMsg {
    let message = target ?? this.create(),
      end = reader.pos + length
    while (reader.pos < end) {
      let [fieldNo, wireType] = reader.tag()
      switch (fieldNo) {
        case /* string msg */ 1:
          message.msg = reader.string()
          break
        case /* string str_data */ 2:
          message.strData = reader.string()
          break
        case /* bytes bin_data */ 3:
          message.binData = reader.bytes()
          break
        default:
          let u = options.readUnknownField
          if (u === 'throw')
            throw new globalThis.Error(
              `Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`
            )
          let d = reader.skip(wireType)
          if (u !== false)
            (u === true ? UnknownFieldHandler.onRead : u)(
              this.typeName,
              message,
              fieldNo,
              wireType,
              d
            )
      }
    }
    return message
  }
  internalBinaryWrite(
    message: ComMsg,
    writer: IBinaryWriter,
    options: BinaryWriteOptions
  ): IBinaryWriter {
    /* string msg = 1; */
    if (message.msg !== '') writer.tag(1, WireType.LengthDelimited).string(message.msg)
    /* string str_data = 2; */
    if (message.strData !== '') writer.tag(2, WireType.LengthDelimited).string(message.strData)
    /* bytes bin_data = 3; */
    if (message.binData.length) writer.tag(3, WireType.LengthDelimited).bytes(message.binData)
    let u = options.writeUnknownFields
    if (u !== false) (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer)
    return writer
  }
}
/**
 * @generated MessageType for protobuf message ld_data.ComMsg
 */
export const ComMsg = new ComMsg$Type()