// applications/logger_fe/src/basic_data/proto/point_cloud.ts
export interface PointCloud {
  /**
   * @generated from protobuf field: string name = 1;
   */
  name: string
  /**
   * @generated from protobuf field: uint64 id = 2;
   */
  id: bigint
  /**
   * @generated from protobuf field: uint64 timestamp = 3;
   */
  timestamp: bigint
  /**
   * @generated from protobuf field: string fields = 4;
   */
  fields: string
  /**
   * @generated from protobuf field: string types = 5;
   */
  types: string
  /**
   * @generated from protobuf field: string shape = 6;
   */
  shape: string
  /**
   * @generated from protobuf field: bool is_bigendian = 7;
   */
  isBigendian: boolean
  /**
   * @generated from protobuf field: bytes data = 8;
   */
  data: Uint8Array
}
// @generated message type with reflection information, may provide speed optimized methods
class PointCloud$Type extends MessageType<PointCloud> {
  constructor() {
    super('ld_data.PointCloud', [
      { no: 1, name: 'name', kind: 'scalar', T: 9 /*ScalarType.STRING*/ },
      { no: 2, name: 'id', kind: 'scalar', T: 4 /*ScalarType.UINT64*/, L: 0 /*LongType.BIGINT*/ },
      {
        no: 3,
        name: 'timestamp',
        kind: 'scalar',
        T: 4 /*ScalarType.UINT64*/,
        L: 0 /*LongType.BIGINT*/,
      },
      { no: 4, name: 'fields', kind: 'scalar', T: 9 /*ScalarType.STRING*/ },
      { no: 5, name: 'types', kind: 'scalar', T: 9 /*ScalarType.STRING*/ },
      { no: 6, name: 'shape', kind: 'scalar', T: 9 /*ScalarType.STRING*/ },
      { no: 7, name: 'is_bigendian', kind: 'scalar', T: 8 /*ScalarType.BOOL*/ },
      { no: 8, name: 'data', kind: 'scalar', T: 12 /*ScalarType.BYTES*/ },
    ])
  }
  create(value?: PartialMessage<PointCloud>): PointCloud {
    const message = {
      name: '',
      id: 0n,
      timestamp: 0n,
      fields: '',
      types: '',
      shape: '',
      isBigendian: false,
      data: new Uint8Array(0),
    }
    globalThis.Object.defineProperty(message, MESSAGE_TYPE, { enumerable: false, value: this })
    if (value !== undefined) reflectionMergePartial<PointCloud>(this, message, value)
    return message
  }
  internalBinaryRead(
    reader: IBinaryReader,
    length: number,
    options: BinaryReadOptions,
    target?: PointCloud
  ): PointCloud {
    let message = target ?? this.create(),
      end = reader.pos + length
    while (reader.pos < end) {
      let [fieldNo, wireType] = reader.tag()
      switch (fieldNo) {
        case /* string name */ 1:
          message.name = reader.string()
          break
        case /* uint64 id */ 2:
          message.id = reader.uint64().toBigInt()
          break
        case /* uint64 timestamp */ 3:
          message.timestamp = reader.uint64().toBigInt()
          break
        case /* string fields */ 4:
          message.fields = reader.string()
          break
        case /* string types */ 5:
          message.types = reader.string()
          break
        case /* string shape */ 6:
          message.shape = reader.string()
          break
        case /* bool is_bigendian */ 7:
          message.isBigendian = reader.bool()
          break
        case /* bytes data */ 8:
          message.data = reader.bytes()
          break
        default:
          let u = options.readUnknownField
          if (u === 'throw')
            throw new globalThis.Error(
              `Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`
            )
          let d = reader.skip(wireType)
          if (u !== false)
            (u === true ? UnknownFieldHandler.onRead : u)(
              this.typeName,
              message,
              fieldNo,
              wireType,
              d
            )
      }
    }
    return message
  }
  internalBinaryWrite(
    message: PointCloud,
    writer: IBinaryWriter,
    options: BinaryWriteOptions
  ): IBinaryWriter {
    /* string name = 1; */
    if (message.name !== '') writer.tag(1, WireType.LengthDelimited).string(message.name)
    /* uint64 id = 2; */
    if (message.id !== 0n) writer.tag(2, WireType.Varint).uint64(message.id)
    /* uint64 timestamp = 3; */
    if (message.timestamp !== 0n) writer.tag(3, WireType.Varint).uint64(message.timestamp)
    /* string fields = 4; */
    if (message.fields !== '') writer.tag(4, WireType.LengthDelimited).string(message.fields)
    /* string types = 5; */
    if (message.types !== '') writer.tag(5, WireType.LengthDelimited).string(message.types)
    /* string shape = 6; */
    if (message.shape !== '') writer.tag(6, WireType.LengthDelimited).string(message.shape)
    /* bool is_bigendian = 7; */
    if (message.isBigendian !== false) writer.tag(7, WireType.Varint).bool(message.isBigendian)
    /* bytes data = 8; */
    if (message.data.length) writer.tag(8, WireType.LengthDelimited).bytes(message.data)
    let u = options.writeUnknownFields
    if (u !== false) (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer)
    return writer
  }
}
/**
 * @generated MessageType for protobuf message ld_data.PointCloud
 */
export const PointCloud = new PointCloud$Type()