// applications/logger_fe/src/basic_data/proto/data_block_idx.ts
export interface DataBlockIdx {
  /**
   * @generated from protobuf field: ld_data.DataBlockType type = 1;
   */
  type: DataBlockType
  /**
   * uint32 type = 1;
   *
   * @generated from protobuf field: uint64 timestamp = 2;
   */
  timestamp: bigint
  /**
   * @generated from protobuf field: uint64 pos = 3;
   */
  pos: bigint // 直接是 protobuf 数据块的起始位置, 不包含头部的起始标识符, 数据块标识符等. 直接从该位置读取数据然后解析.
  /**
   * @generated from protobuf field: uint32 block_size = 4;
   */
  blockSize: number // 每个数据块大小不能超过 4G 字节(2^32).
  /**
   * @generated from protobuf field: string name = 5;
   */
  name: string
}
/**
 * @generated from protobuf message ld_data.DataBlockIndices
 */
export interface DataBlockIndices {
  /**
   * @generated from protobuf field: repeated ld_data.DataBlockIdx indices = 1;
   */
  indices: DataBlockIdx[]
}
/**
 * @generated from protobuf enum ld_data.DataBlockType
 */
export enum DataBlockType {
  /**
   * @generated from protobuf enum value: UNKNOWN = 0;
   */
  UNKNOWN = 0,
  /**
   * @generated from protobuf enum value: FRAME_SPLITER = 1;
   */
  FRAME_SPLITER = 1,
  /**
   * @generated from protobuf enum value: POINT_CLOUD = 2;
   */
  POINT_CLOUD = 2,
  /**
   * @generated from protobuf enum value: OBJECT_LIST = 3;
   */
  OBJECT_LIST = 3,
}
// @generated message type with reflection information, may provide speed optimized methods
class DataBlockIdx$Type extends MessageType<DataBlockIdx> {
  constructor() {
    super('ld_data.DataBlockIdx', [
      { no: 1, name: 'type', kind: 'enum', T: () => ['ld_data.DataBlockType', DataBlockType] },
      {
        no: 2,
        name: 'timestamp',
        kind: 'scalar',
        T: 4 /*ScalarType.UINT64*/,
        L: 0 /*LongType.BIGINT*/,
      },
      { no: 3, name: 'pos', kind: 'scalar', T: 4 /*ScalarType.UINT64*/, L: 0 /*LongType.BIGINT*/ },
      { no: 4, name: 'block_size', kind: 'scalar', T: 13 /*ScalarType.UINT32*/ },
      { no: 5, name: 'name', kind: 'scalar', T: 9 /*ScalarType.STRING*/ },
    ])
  }
  create(value?: PartialMessage<DataBlockIdx>): DataBlockIdx {
    const message = { type: 0, timestamp: 0n, pos: 0n, blockSize: 0, name: '' }
    globalThis.Object.defineProperty(message, MESSAGE_TYPE, { enumerable: false, value: this })
    if (value !== undefined) reflectionMergePartial<DataBlockIdx>(this, message, value)
    return message
  }
  internalBinaryRead(
    reader: IBinaryReader,
    length: number,
    options: BinaryReadOptions,
    target?: DataBlockIdx
  ): DataBlockIdx {
    let message = target ?? this.create(),
      end = reader.pos + length
    while (reader.pos < end) {
      let [fieldNo, wireType] = reader.tag()
      switch (fieldNo) {
        case /* ld_data.DataBlockType type */ 1:
          message.type = reader.int32()
          break
        case /* uint64 timestamp */ 2:
          message.timestamp = reader.uint64().toBigInt()
          break
        case /* uint64 pos */ 3:
          message.pos = reader.uint64().toBigInt()
          break
        case /* uint32 block_size */ 4:
          message.blockSize = reader.uint32()
          break
        case /* string name */ 5:
          message.name = reader.string()
          break
        default:
          let u = options.readUnknownField
          if (u === 'throw')
            throw new globalThis.Error(
              `Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`
            )
          let d = reader.skip(wireType)
          if (u !== false)
            (u === true ? UnknownFieldHandler.onRead : u)(
              this.typeName,
              message,
              fieldNo,
              wireType,
              d
            )
      }
    }
    return message
  }
  internalBinaryWrite(
    message: DataBlockIdx,
    writer: IBinaryWriter,
    options: BinaryWriteOptions
  ): IBinaryWriter {
    /* ld_data.DataBlockType type = 1; */
    if (message.type !== 0) writer.tag(1, WireType.Varint).int32(message.type)
    /* uint64 timestamp = 2; */
    if (message.timestamp !== 0n) writer.tag(2, WireType.Varint).uint64(message.timestamp)
    /* uint64 pos = 3; */
    if (message.pos !== 0n) writer.tag(3, WireType.Varint).uint64(message.pos)
    /* uint32 block_size = 4; */
    if (message.blockSize !== 0) writer.tag(4, WireType.Varint).uint32(message.blockSize)
    /* string name = 5; */
    if (message.name !== '') writer.tag(5, WireType.LengthDelimited).string(message.name)
    let u = options.writeUnknownFields
    if (u !== false) (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer)
    return writer
  }
}
/**
 * @generated MessageType for protobuf message ld_data.DataBlockIdx
 */
export const DataBlockIdx = new DataBlockIdx$Type()
// @generated message type with reflection information, may provide speed optimized methods
class DataBlockIndices$Type extends MessageType<DataBlockIndices> {
  constructor() {
    super('ld_data.DataBlockIndices', [
      {
        no: 1,
        name: 'indices',
        kind: 'message',
        repeat: 1 /*RepeatType.PACKED*/,
        T: () => DataBlockIdx,
      },
    ])
  }
  create(value?: PartialMessage<DataBlockIndices>): DataBlockIndices {
    const message = { indices: [] }
    globalThis.Object.defineProperty(message, MESSAGE_TYPE, { enumerable: false, value: this })
    if (value !== undefined) reflectionMergePartial<DataBlockIndices>(this, message, value)
    return message
  }
  internalBinaryRead(
    reader: IBinaryReader,
    length: number,
    options: BinaryReadOptions,
    target?: DataBlockIndices
  ): DataBlockIndices {
    let message = target ?? this.create(),
      end = reader.pos + length
    while (reader.pos < end) {
      let [fieldNo, wireType] = reader.tag()
      switch (fieldNo) {
        case /* repeated ld_data.DataBlockIdx indices */ 1:
          message.indices.push(DataBlockIdx.internalBinaryRead(reader, reader.uint32(), options))
          break
        default:
          let u = options.readUnknownField
          if (u === 'throw')
            throw new globalThis.Error(
              `Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`
            )
          let d = reader.skip(wireType)
          if (u !== false)
            (u === true ? UnknownFieldHandler.onRead : u)(
              this.typeName,
              message,
              fieldNo,
              wireType,
              d
            )
      }
    }
    return message
  }
  internalBinaryWrite(
    message: DataBlockIndices,
    writer: IBinaryWriter,
    options: BinaryWriteOptions
  ): IBinaryWriter {
    /* repeated ld_data.DataBlockIdx indices = 1; */
    for (let i = 0; i < message.indices.length; i++)
      DataBlockIdx.internalBinaryWrite(
        message.indices[i],
        writer.tag(1, WireType.LengthDelimited).fork(),
        options
      ).join()
    let u = options.writeUnknownFields
    if (u !== false) (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer)
    return writer
  }
}
/**
 * @generated MessageType for protobuf message ld_data.DataBlockIndices
 */
export const DataBlockIndices = new DataBlockIndices$Type()